feat(backup-viewer): Add search
* Moved BackupViewer to it's own folder under components
* Update packages
