+++ package.json
@@ -9,9 +9,9 @@
     "docs:build": "vuepress build docs"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.0",
+    "@types/jest": "^27.4.1",
     "@types/lodash": "^4.14.178",
-    "@types/node": "^17.0.19",
+    "@types/node": "^17.0.21",
     "@vue/component-compiler-utils": "^3.3.0",
     "@vuepress/plugin-pwa": "^2.0.0-beta.35",
     "@vuepress/plugin-register-components": "^2.0.0-beta.35",
